Maybe fix POST(save_labels) 500 (commented out)
Since the user is no longer providing ground truth labels
via the yes button or the drop down button,
I figure the server was mad that no labels were provided (empty list).

Still need to test this out. If it turns out to be the fix,
we'll also need to remove portions of the results page
that require ground truth labels
(i.e. fine tuning, etc from get_results by making a copy to play with).