Can toggle T/F to run finetuning in get_results
Hoped to test my belief that finetuning is what makes get_results
take so long to mount.

So far, that seems to not really be the case, so instead of
removing the function call for finetuning altogether,
I'll be able to switch between the two and see if there is a noticable time difference.

The real culprit of the time seems to be the labeling itself, so consider if there is anything we can do to speed that up or
show the user that it's making progress.

(If we can't, just guide the user to their terminal
which prints out how many unlabeled examples are left
after each labeing iteration).