UI working, but get_final_labels returns []
I was able to run everything without error,
- reformated the Training page so that the
Loading components only show up when necessary
- added a progress bar using tqdm
instead of having print statements for num labeled/unlabeled data.
- also made sure to display what any error was
in my exception for get_final_labels

However, we still need to do maintenance on get_final_labels.
- since I see in the frontend that we only got an empty list.

The labels we tried to pull from the backend seem to either
- not be passing to the frontend correctly
- or not being read from the csv correctly
- or some other problem from the backend function.

So, we'll need to investigate this probably in a test script
outside the app so we don't have to rerun the labeling process
to see if it worked.
- Maybe we can even add some functionality
to not run the labeling if the labels have already been evaluated
(similar to how we don't have to generate a model
when we have one already pretrained,
check the path)