Bring back clean slate NLPDocTool
This way as we move forward, we know the pages we want to have been untouched
won't have been modified from previous experimentation.

Added the state constants for each page and the imports.
Considering placing the openCoding pages in one folder as well.
(Not sure if that would make things clearer or not)

Todo:
- remove routing from NLPDocTool, use opencoding page transitions
- begin outlining step for user to enter their model endpoints and training params
(reference opencoding introduction for what to ask for and how to save
the users preferences there)