All page transitions working
Additional minor changes to prepare for further testing
(i.e. passing in the props to get and post data)

Only problem here is that several pages only show the bottom and
don't allow me to scroll to the top.
I think this is something to do with mui or the theme.
I may just have to do what i did with results and set up grids with row heights.