Added method try to validate user url input. Error
Basically hit a roadblock because I was trying to handle things that could go wrong
with the user's input about their model (i.e. they typed in the url wrong).

But I could not figure out how to catch a timeout error (which is all we would get
if the user entered the wrong url). So, it's now either, they make sure its right
before they enter / submit the info. Or, it has the same error later.

Tomorrow, lets just prompt for the training parameters and try to run it on GPT2
or BART.