Attempted setting multiple proxies in package.json
It errored.
When specified, "proxy" in package.json must be a string.
Instead, the type of "proxy" was "object".
Either remove "proxy" from package.json, or make it a string.

So, I'm going to try again with setupProxy (http-proxy-middleware).
My idea is maybe I wasn't calling the proxy correctly last time.
I think if I stick to having the "url" be the proxy name, it may work.