Attempt to implement /model/predict with pretrain
Note, we'll probably need to do something more like opencoding with training.
But, this is just so I can see how we will maybe go back and forth between
the user's model with get requests.

I'm currently getting this error:
- invalid literal for int() with base 10: 'last_hidden_state'
Which I would guess is because I'm not sure how to decode the model output
(which by itself seems to be a tensor a matrix of floats).

Todo:
- debug /predict decoding (i.e. remove skip_special_tokens, add others)
- add a call to predict in our frontend that we can display to the user
(i.e. so that we can compare the user's hypothesis with the model output).
- consider how we will train a model from scratch on the opencoding results