Begin planning how we'll call /user_model/predict
Added a reminder to changes_README.md that I should start up
the test user model before trying to ping its endpoints.

Added a note about what we need to clarify about the training of the model,
but for times sake I will be focusing on /user_model/predict.

Todo:
- clarify what needs to be done for NLPDocTool backend integration
- maybe comment out the current calls to NLPDocTool process data endpoints,
and just focus on calling predict with dummy data.
- go back to Step5.js and call the user's predict endpoint